switch the mdoc PostModifier to accept almond.Image
