+++ rainier-notebook/src/main/resources/META-INF/services/mdoc.PostModifier
@@ -1,0 +1,1 @@
+com.stripe.rainier.notebook.ImageModifier
