--- rainier-plot/src/main/resources/META-INF/services/mdoc.PostModifier
@@ -1,1 +1,0 @@
-com.stripe.rainier.plot.EvilplotModifier
